Refactoring: extract module data (29-diamond-event)
