+++ proxy/common_neon/data.py
@@ -1,3 +1,10 @@
+# File: proxy/common_neon/data.py
+# Module: data
+# Description: contains simple structs
+
+
+from dataclasses import dataclass, field
+from typing import List
 
 
 class NeonTxStatData:
@@ -10,3 +17,27 @@ class NeonTxStatData:
 
     def add_instruction(self, sol_tx_hash: str, sol_spent: int, steps: int, bpf: int) -> None:
         self.instructions.append((sol_tx_hash, sol_spent, steps, bpf))
+
+
+@dataclass
+class ReturnDTO:
+    exit_status: int = 0
+    gas_used: int = 0
+    return_value: bytes = None
+
+
+@dataclass
+class EventDTO:
+    address: bytes = None
+    count_topics: int = 0
+    topics: List[bytes] = None
+    log_data: bytes = None
+
+
+@dataclass
+class LogIxDTO:
+    return_dto: ReturnDTO = None
+    event_dtos: List[EventDTO] = field(default_factory=list)
+
+    def empty(self) -> bool:
+        return self.return_dto is None
